Refactor: More refactoring
