+++ source/typescript/types/stack_trace_location.ts
@@ -1,0 +1,11 @@
+export type StackTraceLocation = {
+    type: "URL";
+    col: number;
+    line: number;
+    protocol: string;
+    url: string;
+} | {
+    type: "ANONYMOUS";
+    col: number;
+    line: number;
+};
