+++ source/typescript/utilities/create_test_suite.ts
@@ -1,0 +1,13 @@
+import { TestSuite } from "../types/test_suite.js";
+
+
+export function createTestSuite(url_string: string, index: number): TestSuite {
+    return {
+        origin: url_string,
+        tests: [],
+        index,
+        data: "",
+        name: "",
+        url: null
+    };
+}
